script directory as part of the config, header text functions can be from different package/module
