--- pygame_console/requirements.txt
@@ -1,2 +1,0 @@
-pygame>=1.9.4
-
